Refactor `get_object`. Add version_id to `S3_File`. Add support for versions on get and head object.
